% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{create_path}
\alias{create_path}
\title{Create coordinates following defined trajectory}
\usage{
create_path(
  n_points = 100,
  nodes = c(1, -1, -1, 2, 0),
  node_int = NULL,
  scale = 1
)
}
\arguments{
\item{n_points}{An integer. Controls output "resolution". (Underlying calculus is unaffected).}

\item{nodes}{A numeric vector corresponding to 2nd derivative values. This determines the overall shape of the function.}

\item{node_int}{A numeric vector assigning x-values for nodes. Automatically calculated if \code{NULL}.}

\item{scale}{A number. Adjusts all y-values so that max(y) = \code{scale}.}
}
\value{
A tibble containing x & y coordinates of resulting function.
}
\description{
Manually create data points corresponding to piecewise polynomial.

Use \code{create_path()} to exert full control over trajectory shape. Starting
with scalar values defined by \code{nodes}, \code{create_path()} integrates twice to
create continuously differentiable function.

\code{create_path()} is the underlying function to \code{meandr()}.
}
