% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{Create simply line plot from x + y coordinates. Designed for `meandr` but can be used in other applications.}
\usage{
mplot(x)
}
\arguments{
\item{x}{A numeric vector or data frame containing at least two numeric columns.
Numeric vector will be automatically paired with x-values in sequence from 0 to 1.}
}
\value{
An object of class `gg`
}
\description{
Create simply line plot from x + y coordinates. Designed for `meandr` but can be used in other applications.
}
