% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{natural}
\alias{natural}
\title{Force numeric to natural number}
\usage{
natural(x, x_arg = "x")
}
\arguments{
\item{x}{A numeric}

\item{x_arg}{Argument name for x. Used in error message to inform user about location of incompatible type}
}
\value{
integer
}
\description{
Force numeric to natural number
}
