% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quasi_equal}
\alias{quasi_equal}
\title{Check if two numbers are approximately equal}
\usage{
quasi_equal(x, y, tol = 1e-07)
}
\arguments{
\item{x}{A number}

\item{y}{A number}

\item{tol}{A number}
}
\value{
logical
}
\description{
Check if two numbers are approximately equal
}
