% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quasi_greater_equal}
\alias{quasi_greater_equal}
\title{Check if one number is approximately greather than or equal to another}
\usage{
quasi_greater_equal(x, y, tol = 1e-07)
}
\arguments{
\item{x}{A number}

\item{y}{A number}

\item{tol}{A number}
}
\value{
logical
}
\description{
Check if one number is approximately greather than or equal to another
}
