% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_roll}
\alias{sample_roll}
\title{Perform sequential sample with weighted replacement}
\usage{
sample_roll(x, n, gain = 0.75)
}
\arguments{
\item{x}{vector of elements from which to choose}

\item{n}{a positive number, the number of items to choose from}

\item{gain}{Tuning parameter.}
}
\value{
numeric vector
}
\description{
Perform sequential sample with weighted replacement
}
