% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandr.R
\name{meandr}
\alias{meandr}
\title{Simulate random coordinates following a continuous differentiable trajectory.}
\usage{
meandr(
  n_points = 100L,
  n_nodes = 100L,
  wt = c(1, -1),
  scale = 1,
  seed = NULL,
  output = c("vector", "tibble")
)
}
\arguments{
\item{n_points}{An integer. Controls output "resolution". (Underlying calculus is unaffected).}

\item{n_nodes}{An integer. Defines number of distinct inflection points in function.}

\item{wt}{A numeric vector of values. These will be sampled (with replacement) \code{n_nodes} times to create inflection points.}

\item{scale}{A number. Adjusts all y-values so that \code{max(y) = scale}.}

\item{seed}{A number passed to \code{set.seed} for repeatability. If \code{NULL}, no seed created.}

\item{output}{Either \code{"vector"} or \code{"tibble"}.}
}
\value{
A tibble containing x & y coordinates of resulting function.
}
\description{
meandr
}
\details{
recommended method for wt
}
