% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{Create simply line plot from x + y coordinates. Designed for `meandr` but can be used in other applications.}
\usage{
mplot(.df)
}
\arguments{
\item{.df}{A data frame containing at least two numeric columns.}
}
\value{
An object of class `gg`
}
\description{
Create simply line plot from x + y coordinates. Designed for `meandr` but can be used in other applications.
}
\examples{
mplot(data.frame(x = 1:10, y = 1:10))

mplot(meandr())
}
